Corregir el tipo de conversión de balance de flotante a entero en el formulario de liquidación

+++ src/transportation/pages/control-transport-detail/index.tsx
@@ -1246,7 +1246,7 @@ const LiquidationBalance = () => {
               type="number"
               value={balance.balance}
               onChange={(e) =>
-                setBalance({ ...balance, balance: parseFloat(e.target.value) })
+                setBalance({ ...balance, balance: parseInt(e.target.value) })
               }
               required
             />
